update of test expected results.
update of test expected results. They were slightly different as a result of the ordering of the dijkstra queue becoming strict instead of weak.
Hopefully, the result will now be deterministic independantly of the platform. Tests result used to be different on 32 bit and 64 bit platforms, probably because of the weak ordering of the dijkstra queue. Future testing will prove me right or wrong.