rectangle frame computation C++/JS harmonisation
rectangle frame is computed by bombix in C++ the first time.
When a box is dragged and released, the rectangle frame is computed again to take the one rectangle drag into account. This time it gets computed in diagload.js in Javascript. Both codes need to be harmonised. Before this commit, it was not the case, there was a slight difference, which gave the impression of instability of the rectangles on the screen after the first drag + release user interaction.